third_party/blink/web_tests/platform/generic/fast/frames/frame-set-location-badstring-expected.txt



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
Tests that assigning a bad string to HTMLFrameElement.location throws and aborts properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS frameLocation is not null
PASS frameLocation is not undefined
PASS parent.frames[1].location = badString; threw exception Exception in toString().
PASS parent.frames[1].location is frameLocation
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: 'r'
--------
